requirements.txt
python-telegram-bot==20.6
apscheduler==3.10.4
pytz==2023.3
pandas==2.0.0